WIP:
Building create form - not exactly sure how to best handle data array

What?
(1) useRef to obtain the values of each input to package into payload to be sent to firebase db

Problems?
(1) easy to manage the state of 1 singular input - using ref
(2) not sure how to best handle multiple refs - or nested layers? Apparently Redux and other tools are good for this?